+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Changelogs for this project are recorded in this file since v0.2.0.
 
 ### Added
 
-* `dtw_dist_from_metric` allows one to pick a dedicated ground metric on top
+* `dtw_from_metric` allows one to pick a dedicated ground metric on top
 of which the DTW algorithm can be run but only return the similarity 
 
 ## [v0.5.3]